Update all .po files with changes in .pot

po_update() updates existing .po file after the .pot file has changed. There are four cases:

New messages: added with blank msgstr.

Deleted messages: marked as deprecated and moved to the bottom of the file.

Major changes to existing messages: appear as an addition and a deletion.

Minor changes to existing messages: will be flagged as fuzzy.

#, fuzzy, c-format
#| msgid "Generating en@quot translations"
msgid "Updating '%s' %s translation"
msgstr "memperbarui terjemahan bahasa en@quot..."

The previous message is given in comments starting with #|. Translators need to update the actual (uncommented) msgstr manually, using the old msgid as a potential reference, then delete the old translation and the fuzzy comment (c-format should remain, if present).

Usage

po_update(dir = ".", lazy = TRUE, verbose = !is_testing())

Arguments

dir

Character, default the present directory; a directory in which an R package is stored.

lazy

If TRUE, only .po files that are older than their corresponding .pot file will be updated.

verbose

Logical, default TRUE (except during testing). Should extra information about progress, etc. be reported?